Add LinkSource as input type.
Instead of taking the content as an AsRef<str>, now I gather file origin
metadata for the links. This is useful in two regards: first and foremost, it
answers the debugging question, "where did this link come from?" Secondarily,
it lets me use the date-stamp in the filename added by my note-taking scripts.

There's still some waste here, in that the filename gets turned to a String
about a million times; it's worth looking into whether or not SQLX will accept
a Cow instead of a string in this case.

Thanks to Kat Marchán for suggesting a much less heinous way to use `let/else`
in this implementation [1].

[1]: https://hachyderm.io/@zkat@toot.cat/109397055825407654